feat : 부분 공개를 위한 그룹 권한 추가 및 수정 기능
부분 공개를 할 유저 추가 전 단계
-  group_level : 0, 10, 11을 설정해서 그 게시글에 대한 추가 권한 부여
- 게시글이 비공개(prelevel = 0) 되어있다면 그룹 권한을 못하게 하는 예외처리 필요
#27

+++ App/src/main/java/com/mylittlelib/app/model/AuthorityGroup.java
@@ -19,7 +19,7 @@ public class AuthorityGroup {
     private Long groupIndex;
 
     @Column(name = "group_level")
-    private Long groupLevel;
+    private int groupLevel;
 
     @OneToOne
     @JoinColumn(name = "category_index")
